requirements.txt
celery==4.3.0
codecov==2.0.15
Django==1.11.6
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-debug-toolbar==1.8
djangorestframework==3.9.2
django-rest-swagger==2.2.0
djangorestframework-simplejwt==4.3.0
django-extensions==2.1.6
django-filter==2.1.0
django-haystack==2.8.1
django-structlog==1.2.3
elasticsearch==2.4.1
google-cloud-pubsub==0.39.1
gunicorn==19.9.0
locustio==0.11.0
mysqlclient==1.4.2
Pillow==6.0.0
pytest==4.4.0
pytest-cov==2.6.1
pytest-django==3.4.8
python-logstash==0.4.6
redis==3.2.1
whitenoise==4.1.2
